feat: adapt rocm_sdk_devel build for v26.4
